backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.1
pytest==7.4.2
requests==2.31.0 